Add files via upload
This is the final assignment of Applied Data science Capstone.

The problem:

A security technology company wants to offer a new surveillance device to commercials in Toronto. The company wants to focus on commercial located in neighborhoods that have presented breaking and entering in 2019, that information is part of tha market study. The study need information from the last year in order to know the neighborhoods where occur that kind of crime.

We will use data from Toronto Police Department and Foursquare.

Enjoy it!